requirements.txt
google-generativeai
python-telegram-bot
python-dotenv